Updated a working tutorial 2 and back added a function name update to prior tutorial FaradayIO's
