+++ Tutorials/FaradayIO/Tutorial_2/FaradayIO/deviceconfig.pyc
