+++ Tutorials/FaradayIO/Tutorial_2/FaradayIO/faradaycommands.pyc
