+++ Tutorials/FaradayIO/Tutorial_2/FaradayIO/gpioallocations.pyc
